requirements.txt
pypdf
chromadb
google-generativeai
python-dotenv 